[feat]: merge ViewEnvironmentPropagatingObject into ViewEnvironmentPropagating (#218)
